// frontend/src/shared/types/model/webIncident.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sentinel Monitoring API
 * API for service monitoring and incident management
 * OpenAPI spec version: 1.0
 */

/**
 * Service incident
 */
export interface WebIncident {
  duration?: string;
  id?: string;
  message?: string;
  resolved?: boolean;
  resolved_at?: string;
  service_id?: string;
  service_name?: string;
  started_at?: string;
  status?: string;
}